src/cores/cicada-ingestion/data-providers/Binance/sources/ingestion-data-vision-binance/requirements.txt
pandas
protobuf
python-dotenv
Requests
tqdm
urllib3
google-cloud-storage
schedule
pyarrow
